requirements.txt
apparmor==3.0.1
appdirs==1.4.4
application-utility==1.3.2
asn1crypto==1.4.0
astroid==2.5.1
attrs==20.3.0
beautifulsoup4==4.9.3
bleach==3.3.0
bs4==0.0.1
btrfsutil==5.11
CacheControl==0.12.6
catfish==4.16.0
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.4.6
cssselect==1.1.0
distlib==0.3.1
distro==1.5.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
fake-useragent==0.1.11
gufw==21.4.0
html5lib==1.1
idna==2.10
iniconfig==1.1.1
isort==5.7.0
jeepney==0.6.0
keyring==22.0.1
keyutils==0.6
lazy-object-proxy==1.5.2
lensfun==0.3.95
LibAppArmor==3.0.1
lightdm-gtk-greeter-settings==1.2.2
louis==3.17.0
lxml==4.6.2
lyricwikia==0.1.11
Markdown==3.3.4
mccabe==0.6.1
menulibre==2.2.1
mps-youtube==0.2.8
msgpack==1.0.2
mugshot==0.4.3
netsnmp-python==1.0a1
nftables==0.1
npyscreen==4.10.5
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.20.1
pafy==0.5.5
paho-mqtt==1.5.1
parse==1.19.0
pep517==0.9.1
pexpect==4.8.0
Pillow==8.1.0
pkginfo==1.7.0
pluggy==0.13.1
ply==3.11
progress==1.5
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.0
pycparser==2.20
pyee==8.1.0
Pygments==2.8.0
PyGObject==3.38.0
PyJWT==1.7.1
pylint==2.7.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyppeteer==0.2.5
PyQt5==5.15.4
PyQt5-sip==12.8.1
pyquery==1.4.3
pysmbc==1.0.23
pytest==6.2.2
pytest-runner==5.3.0
python-distutils-extra==2.39
pytz==2021.1
pyxdg==0.26
PyYAML==5.4.1
rados==2.0.0
rbd==2.0.0
readme-renderer==29.0
reportlab==3.5.65
requests==2.25.1
requests-html==0.10.0
requests-toolbelt==0.9.1
resolvelib==0.5.4
retrying==1.3.3
rfc3986==1.4.0
rgw==2.0.0
SecretStorage==3.3.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2
spotify-cli-linux==1.6.0
team==1.0
thingooConnector @ file:///home/ninjavaz/Desktop/connector-python/dist/thingooConnector-0.0.1-py3-none-any.whl
toml==0.10.2
tqdm==4.57.0
twilio==6.54.0
twine==3.3.0
udiskie==2.3.3
ufw==0.36
urllib3==1.26.3
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
wrapt==1.12.1
youtube-dl==2021.3.14
